Fixed and reorganized a little the code
